refactor: convert patch script from .mjs to .ts and export patch functions
- Use TypeScript with proper types for patch functions
- Export patch functions so tests import real implementation instead of duplicating regex logic
- Use tsx to run postinstall script, consistent with other scripts
